Revert most ConsoleOutputTransmitterPerformanceTest changes
Something is working even more unreliably on Windows, so perhaps I have a misunderstanding about the code or there is an existing Windows bug. Let's try the old methodology but with a lower interval.

+++ common/src/main/java/com/thoughtworks/go/remote/work/ConsoleOutputTransmitter.java
@@ -43,7 +43,7 @@ public final class ConsoleOutputTransmitter implements TaggedStreamConsumer, Run
         this(consoleAppender, new SystemEnvironment().getConsolePublishInterval(), TimeUnit.SECONDS, new ScheduledThreadPoolExecutor(1));
     }
 
-    ConsoleOutputTransmitter(ConsoleAppender consoleAppender, Integer consolePublishInterval, TimeUnit consumePublishIntervalUnit, ScheduledThreadPoolExecutor scheduledThreadPoolExecutor) {
+    ConsoleOutputTransmitter(ConsoleAppender consoleAppender, long consolePublishInterval, TimeUnit consumePublishIntervalUnit, ScheduledThreadPoolExecutor scheduledThreadPoolExecutor) {
         this.consoleAppender = consoleAppender;
         this.executor = scheduledThreadPoolExecutor;
         executor.scheduleAtFixedRate(this, 0L, consolePublishInterval, consumePublishIntervalUnit);
